requirements.txt
python-telegram-bot==13.15
Pillow==10.2.0